+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Feature: Add `#originate` method to console as alias for `Adhearsion::OutboundCall.originate`
   * Bugfix: Removed unnecessary Mocha reference from generated plugin
   * Bugfix: Call loggers should be deleted after a call finishes
+  * Bugfix: A menu definition's block context is now available
 
 # [2.3.5](https://github.com/adhearsion/adhearsion/compare/v2.3.4...v2.3.5) - [2013-06-06](https://rubygems.org/gems/adhearsion/versions/2.3.5)
   * Bugfix: Fix race conditions in barging in before output start is acknowledged